+++ CHANGELOG.md
@@ -5,3 +5,5 @@ This mod is distributed under Version 2 of the GPL.  A copy of the GPL is includ
 [Changelog history](https://github.com/Hoizame/AtlasLootClassic/blob/master/AtlasLootClassic/Documentation/Release_Notes.md)
 
 ## v1.4.4 (Dec. XX, 2019)
+
+- Favourites lists are now sorted by name